Documentation fixups.
The local changes file is moved to CHANGES.
The README.md is updated to address some inaccuracies and provide
a little more detailed information, also ensuring that correct
credit and licensing information is given.

Since we don't distributed under GPLv3, remove the file because it
was confusing.

Likewise, the NEWS file was misleading -- NEWS simply *wasn't*.