fixes #29 memory leak
This does not fix all "leaks" that may be detected.  In particular
during program initialization copies of a few strings and such are made
and those copies may persist.  However, those are benign as the program
does not continuously leak memory.  Furthermore, fixing all of them would
be substantially more effort than its worth -- in fact the code to "fix"
this would grow the text size of the program by more than the couple of bytes
of leaked memory!