Pass errret to setupterm
The setupterm(3) manual states:

  If errret is not null, then setupterm returns OK or ERR and stores a
  status value in the integer pointed to by errret.
  ...
  If errret is null, setupterm prints an error message upon finding an
  error and exits.

Since less(1) passes NULL, it will not return on error but instead
terminate the program.

Checking the returned value against ERR requires including curses.h
which causes a naming conflict with the existing beep function.
Therefore kept the condition as is.

The error message is borrowed from setupterm when errret is NULL and
does currently not distinguish between error code 0 and -1.