Fix segfault on inverted search
How to reproduce:

$ cat in
foo
bar
bar
$ less -g in # type: /!foo^J

When performing an inverted search in less, make sure to invalidate the
match bounds prior calling regexec(). In this inverted scenario a match
is found when regexec() returns false causing the bounds to not be
updated. This is problematic since the bounds will then refer to a
previous match and future pointer arithmetic will eventually be off
which is manifested in a SIGSEGV.

Issue reported by Larry Hynes on the OpenBSD tech mailing list[1].

[1] http://marc.info/?l=openbsd-tech&m=150114947021801&w=2